test(android): split the tracker tests where the file went over budget
`AuthoredTokenTrackerTest` reached 162 lines against a limit of 150 once the digit
cases joined the context cases already there. The context tests are their own
subject — which separators keep the word a prefix follows, and what counts as a
word at all — and they move to a file named for it.

The new file says out loud that it reads alongside the iOS
`PersonalSuggestionTokenTests`, since the divergence this branch fixes is exactly
what happens when the two stop describing the same keyboard.

Nothing caught the overrun: no workflow runs `check-kotlin-loc.sh`, and none
builds or tests the Android module at all.
